+++ src/models/group.js
@@ -4,22 +4,12 @@ import User from './user';
 
 const modelName = 'Group';
 
-const imageSchema = new mongoose.Schema({
-  // This field will always be empty in the database
-  // It is just a field that will be used manually
-  URL: String,
-  caption: { type: String, default: '' },
-  fileName: { type: String, required: true, unique: true },
-  creator: { type: mongoose.Schema.Types.ObjectId, ref: User, required: true },
-  dateUploaded: { type: Date, required: true, default: new Date() },
-});
-
 const groupSchema = new mongoose.Schema({
   inviteCode: { type: String, required: true, unique: true },
   users: { type: [{ type: mongoose.Schema.Types.ObjectId, ref: User }], default: [] },
   creator: { type: mongoose.Schema.Types.ObjectId, ref: User, required: true },
   invitedUsers: { type: [{ type: mongoose.Schema.Types.ObjectId, ref: User }], default: [] },
-  images: { type: [imageSchema], default: [] },
+  images: { type: [mongoose.Schema.Types.ObjectId], ref: 'Image', default: [] },
   publicGroup: { type: Boolean, default: false },
   name: { type: String, required: true, default: 'New Group' },
 });
@@ -46,7 +36,4 @@ groupSchema
 
 const Group = mongoose.model(modelName, groupSchema);
 
-const Image = mongoose.model('Image', imageSchema);
-
 export default Group;
-export { Image };
